% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNum.R
\name{formatNum}
\alias{formatNum}
\title{Format a number}
\usage{
formatNum(x = NULL, symbol = "", digits = 1)
}
\arguments{
\item{x}{a numeric value}

\item{symbol}{a symbol like R$}

\item{digits}{digits to considering (Default: 1)}
}
\description{
Let numbers readable for humans.
}
\examples{

library(ggplot2)
ggplot(cars, aes(x = dist, y = speed)) +
 geom_point() +
 scale_y_continuous(label = formatNum)

formatNum(seq(100000, 1000000, 100000))

}
