% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg.R
\name{seg}
\alias{seg}
\title{Split in Equal Groups}
\usage{
seg(n, k)
}
\arguments{
\item{n}{size of vector}

\item{k}{number of groups}
}
\value{
a numbered vector of k groups.
}
\description{
Split a vector of size n into k groups
}
\examples{

x <- c(1:100)
sge(length(x), 3)
split(x, sge(length(x), 3))

}
