% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium.R
\name{selenium}
\alias{selenium}
\title{Selenium Server}
\usage{
selenium(x, ...)
}
\arguments{
\item{x}{start or end}

\item{...}{See sections}
}
\description{
Selenium Server
}
\section{start}{

port: an integer between 4445 and 4490

port2: an integer between 5901 and 5946

auto: detect the ports that are not being used and start the selenium server
with it

version: the version of selenium (Default: 3.141.59-zinc)

browser: firefox or chrome
}

\section{end}{

port: an integer between 4445 and 4490

port2: an integer between 5901 and 5946
}

