% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitGroupEqual.R
\name{splitGroupEqual}
\alias{splitGroupEqual}
\title{Split Group Equal}
\usage{
splitGroupEqual(n, k)
}
\arguments{
\item{n}{total number of observations}

\item{k}{number of groups}
}
\value{
a numbered vector of k groups.

uKGroup

uKGroup(100, 2)

uKGroup(100, 3)
}
\description{
This function is useful to split a data into
k equals groups.
}
