% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNum.R
\name{formatNum}
\alias{formatNum}
\title{Format a number}
\usage{
formatNum(x = NULL, smbl = "", signif = 1)
}
\arguments{
\item{x}{a numeric value}

\item{smbl}{a symbol like R$}

\item{signif}{digits to considering (Default: 1)}
}
\description{
Let numbers readble for humans.
}
\examples{

library(ggplot2)
ggplot(cars, aes(x = dist, y = speed)) +
 geom_point() +
 scale_y_continuous(label = gFormatNum)

formatNum(seq(100000, 1000000, 100000))

}
